1. IMPORTANT CHANGE for all code versions: Qinter was not being defined before it is used in irradiance_sog so for the first time step it was uninitialized and after that it used the value from the previous time step.
Moved Qinter and Einter code up in the loop to just before
irradiance_sog is called.  This moves it into the part of the code
only done the first time through the loop which is fine... these
values are dependent only on time and not on water column state
variables.

2. Following critical change in kcollins06. Surface salinity parameter
used in call to surface_flux_sog changed from S%new(0) to S%old(1)

3. Following critical change in kcollins06. Under convection, beta_t
   was being calculated based ratio of minimum value of vertical
   buoyancy flux to surface flux.  This is highly unstable and not as
   suggested in Large et al.  Changed so that under convection beta_t
   = -0.2.  Under stable circumstances beta_t is zero (as it was
   previously).

4. Following important change in kcollins06. dt added to call to
   surface_flux_sog.