Added flagellates_on logical flag line to top of file.
Changed variable descriptions to make them more informative.

The code now reads the data in this file from stdin, so run the code
as:

$ ./SOG < infile

That means that infile is really just a template/sample, and input
files with other names can be created and fed to the code with input
redirection.  All run parameters will eventually be read the same way.