Added derivs_noflag to EXTERNAL declaration.  This is a temporary work-around for problems in the flagellates part of the model.
Added declaration of flagellates flag.  .true. means flagellates model
is included in code.  Added IF block around odeint CALL to use
derivs_sog when flagellates is .true., and derivs_noflag when
flagellates is .false.  This is a temporary work-around for problems
in the flagellates part of the model.

flagellates flag value, inputfile name, and profile output files names
are now read from stdin by redirection of a file formatted like
infile.