As only the length of the grid is used, send in just M not the whole grid variable.
Send in PZ_bins rather than the length M2

As PZ is sent in, it is split into quantities (mico, nitrate etc) and detritus.
Make this more obvious by calling first part PZ_quant

Change the array size on PZ_quant to reflect that it is only the quantities
not the whole array.  NOTE this is a substantive change but had no effect
on the results when running on coho.

Use PZ%bins rather than hard coding the position of quantities.

White space clean up