Refactored.
Wrapped the subroutine in a module.

Changed the argument order to put intent(in) 1st, then intent(out).

Changed the argument arrays to be assumed-shape arrays.

Added the Coriolis factor (f) as an incoming argument.  That
eliminates the need for the module to USE surface_forcing.

Changed the argument names to match (to the extent
reasonable/possible) the names used in the CALL and/or to make them
more descriptive.

Vectorized the Gvector_c calculation.  That eliminates the do-loop and
the local loop index variable, and eliminates the need for the grid
size in the subroutine.  That, in turn, allows the mm argument to be
deleted, and the USE of mean_params to be eliminated.

Added USE precision_defs, and changed double precision argument
declarations to real(kind=dp).

Results of the nov04-test run were not significantly changed (a single
salinity value in the time series output changed by 0.0001) by this
refactoring.