New module.
Provides alloc_check and dalloc_check subroutines for other modules to
use to check allocation/deallocation status, and print an error
message in the case of failure.

May be used in future as the place for allocation of memory for items
that don't logically fit in any other module (but only as a last
resort).