Changed cruise_id.  No longer read from the inputfile, it is not getpars from the infile and is now a 4 character string matching Rich's code for the cruises.
Changed call to initial_mean to include Silicon and specify which parts
of the N% variable are actually being set (eg N%O%new and N%H%new).  This
should probably be done for all the variables.

These changes have no effect on the results except the first line of the
timeseries biology data.