This subroutine needs a good cleaning.  I just cleaned nitrate and added silicon.
Removed constants for nitrate for the Large profile and removed the
Large profile for Nitrate (code was calculating this and then
over-writing it with the data)

N%O%new and N%H%new are now arguments (as NO and NH) rather than all of N.
Sil%O%new an argument.

Code uses nutrient files found in sog-iniital (NOTE spelling mistake).  You
will need to check out this project.

Code chooses nutrient file based on cruise_id.  Opens it, reads the
nitrate and silicon data and checks that the depth is as expected.