Changed the calls to diffusion.  The old diffusion did three things: 1) Calculate the Bmatrix to do classic K diffusion later 2) Calculate nonlocal fluxes and distributed fluxes (Temp & Sal only) 3) Calculate the surface and deep fluxes
These latter two are put in the Gvector which is initialize here.

I separated (1) into a separate routine diffusion_coeff
and (2) and (3) are done for T&S by diffusion_nonlocal_fluxes
and just (3) is done for U&V and biology by diffusion_bot_surf_flux

The old code was using the same bottom flux for NH, Pmicro, Pnano and
Detritus.  The new code fixes this.  This change thus changes the results
in the profiles.

********

Also removed extraneous comments and deleted commented out write statements.