find_upwell is now a module including two subroutines:
upwell_profile (the old find_upwell)
vertical_advection (the old define_adv_bio)

upwell_profile is much re-written to use real (double precision) arithemetic
and interpolation rather than integer arithemetic.  It is now written with
care considering quantities on grid points and on interfaces.  It is written
for un-even grids.  The two parameters 68% of fwc and 2.5 * d have been
updated to the final values found from the data (see entrainment.pdf).

vertical_advection is re-written for clarity but for even grids is unchanged
in mathematics.  The intermediate value P_bio is no longer produced and sent
out of the subroutine as it is not used.  Note that this allows the removal
of variables P_sa, P_no etc.

The changes in upwell_profile do change the results and so you will need
to re-run your reference cases.