Module containing the Runge-Kutta solver for the biological model.
Updated from the three fortran subroutines to be written in f90 and
to use variable names that make sense in the context of SOG

Carries through M, Temperature and I_par.

Hard-coded throughout to use derivs_sog (global search and replace to
change this subroutine name)

Removed un-used intermediate step write-out.

There is an ERROR in one line of one subroutine which I have not yet fixed
as I want to make sure this encapsulation into a module makes no difference
to the code.

This module replaces odeint.f, rkck.f and rkqs.f and they can all be sent
to the attic.