New time series output module.  This module replaces write_open.f90 and write_physical.f90.
3 subroutines:

timeseries_output_open reads the file names for the output files,
opens them, and writes their headers.

write_timeseries is call at each timestep to write results to the
output files.

timeseries_output_close closes the output files at the end of the run.

4 time series output files:

std_phys is a collection of standard physical quantities that everyone
probably needs

user_phys is the place to write quantities for debugging, or of
temporary, or special interest to you.

std_bio and user_bio serve the same purpose for biological quantities.

Changes to what is written to the user_* files should not generally be
committed to the project repository.  The std_* files output should
not generally be changed without good reason.

Note that the file headers and their contents must be kept in sync to
avoid breaking compareSOG plotting scripts.

Comments in the module reiterate all of this.

write_open.f90 and write_physical.f90 can be sent to the CVS attic.

The infile template is changed significantly by the introduction of
this module.  You will need to update any input files you have derived
from the template.