Changed the surface fresh water flux algorithm in two ways.
1. The amount of fresh water added is enough to make the whole mixing layer
have the parameterized surface salinity (not just the top grid level).  This
change is actually made in SOG where the dummy argument passed to dtdz
becomes dt/h%new

2. The algorithm is changed.  A goal for the surface salinity is calculated
based on the previous time step salinity, the parameterization and smearing
the change over half a day. (like previous).

The amount of fresh water flux is incremented (or decremented) as opposed
to set, to get closer to the salinity goal.  This incrementation or
decrementation is necessary because it probably takes some fresh water
flux just to keep the salinity where it is!

The change in fresh water flux is decreased by the iteration count. This
is because there is an inherent oscillation between the mixed layer depth
and the fresh water flux.  (If this surface flux algorithm continues to
work.. we probably should change the iteration solver so that the surface
salinity is set, as opposed to using a fresh water flux. However, if it
proves necessary to distribute the fresh water flux through the upper layers
of the ocean, we will want to stay with an Ft algorithm.)