Replaced Cp_profile() function with calc_rho_alpha_beta_Cp_profiles() subroutine, which will eventually calculate the profiles of all 4 of the required water properties.  Doing them all in 1 subroutine reduces the number of times we need to compute the square root of the salinity profile.
Moved declaration of Cp in from SOG and made it a public variable for
SOG to use.  This simplifies the memory allocation and deallocation
subroutine calls.  Changed arguments of alloc_water_props() and
dalloc_water_props() accordingly.

Made water_property type definition private to the module.

Started adding density (rho) to module, but this is not complete.