Added a new interpolation function, interp_i_d(), to return the value at a specified depth of a quantity stored at the grid layer interfaces.
Note that there is a bug in interp_i_d() that arises when d is close
to zero.  The function should not be used until that bug is fixed.

Changed the name of interp_d() to interp_g_d() because we now have
depth interpolation functions for both grid point and interface
quantities.

Changed the output of warning messages to appear on stdout instead of
stderr.