Added Fw_scale as an incoming argument.  Changed Ft from an inout to an outgoing argument.
Removed fresh water contribution to w%s(0) because that is now
included by way of the surface buoyancy forcing (Bf) calculated in
buoyancy.f90.  *** But we need to check the implications of w%s(0) = 0
on def_gamma.f90. ***

Added some comments around the surface flux code, mostly referring to
equations in Large, etal (1991).