Refactored the buoyancy calculations, and the initialization part of the implicit solver loop.
Added rho to the only clause of the use water_properties statement so
that its %i component can be used for the non turbulent heat flux
calculation.  Added instances of interp_i() to calculate rho%i from
density%new in the 2 places where whater properties are calculated.
Moved call to coefficients to be with the rest of the water properties
code.  Note that refactoring of the water properties is not yet
complete; this is a step along the way.

Added inter_i_d() to only clause of use grid_mod statement, and
changed named of interp_g() to interp_g_d() to reflect changes in grid
module.  Note that interp_i_d() is not ready for use yet.

Moved declaration of Ft into declarations module with the rest of the
fresh water variable declarations, and eliminated its initialization
to zero.  Added code to read Fw_scale, and Fw_depth from the infile.

Added code to read the maximum iteration count for the implicit solver
loop from the infile.

Moved a bunch of initialization code, including calls to find_wind,
and irradiance_sog, out of the top of the implicit solver loop, but
still inside the time step loop.  This was possible because none of
the quantities involved depend on water column values that are changed
in the implicit loop.

Changed the order of execution near the top of the implicit solver
loop.  surface_flux_sog is called 1st to calculate various surface
flux quantities.  Then the nonturbulent heat and fresh water fluxes
are calculated.  Those calculations probably belong in a separate
subroutine in a yet to be created, new surface_forcing module
(sifferent from the existing misnamed one).  Next, buoyancy is called
to calculate the buoyancy profile, and the surface buoyancy forcing.
After that, the implicit solver loop is essentially unchanged (for
now...).

Changed all instances of Bf%b(0) to Bf, which is now a scalar
variable, because it is defined as a surface quantity, and none of the
other array elements were being used.