Added new detritus profile arrays variable (D) to use core_variables statement only clause.
Changed arguments in initial_mean() call to those for refactored
detritus.

Added temporary "bridge code" to transform refactored detritus arrays
to old form so that the refactoring can be done, tested, and committed
in small pieces.  The bridge code will move around as the refactoring
proceeds, and ultimately disappear.

Changed all instances of Detritus(i)%D%old to Detritus(i)%D%new
because the %old and %new values are always the same, so the %old
component has been eliminated.