Changes related to renaming and restructuring of biology modules and subroutines.
Added a new subroutine read_sink_params() to read sinking rate
parameter values from the infile.  This subroutine is called by
init_biology() in the biology_model module.

With the read-in sinking parameter values, the mess above the calls to
sinking_advection() was largely cleaned up.  Added an *** note about
the fact that micro%Nlimit is a profile, but sinking_advection() can
only accept a scalar sinking rate.  Added a call to
sinking_advection() for refractory detritus.