8/10
Added a flag microzooplankton to determine if the microzooplankton do anything.

Added Z to all variable structures etc.  But not that they do not do anything
yet.  Their derivatives are set equal to zero (so this set of changes should
diff clean with the previous except where Z has been added to the output).