New module to hold code for user-specified profiles and time series output (in contrast to the results output in the standard files).
This module may be freely modified in user working copy directories
without checking it into CVS.  In fact, changes should *not* be
checked in.

The goal of the code in this module is to provide a way of getting
exploratory, debugging, or other output of temporary interest without
modifying the standard output that the compareSOG package depends on,
and without having to remember to remove your personal changes from
other modules before you check them in.