Added a new target: install.
make install build the code with development compiler flags, and
creates the necessary directories for running the code (infiles,
outfiles, profiles, and timeseries).

This is useful for setting up a working copy directory after a new
checkout.  I've added it now to sync up with the README.txt file in
the testRunner package, but I should have added it long ago.