Changes related to mixing layer depth code refactoring:
Deleted the initial mixing layer depth parameter value declarations
(hm) because the mixing layer depth is now initialized in
init_mixing_layer().

Renamed mixing layer depth argument from hi to hml, and changed it
from intent(out) to intent(in).  It is used in a depth test for
initialization of some nutrients.