Changed to the new, better than ever parameterization for the light extinguishment constant KK...
Okay real change is that river flow is coming in as the square.  If you just
plot versus the river that's what it looks like.  Also a constant that is
surface intensified is used.

Also changed to correctly use i_space instead of g_space in multiplying the
light downward.

Changing KK means the diff is not clean.