1. Removed (Q/Qbar)**0.41 from the calculation of Ft in not yet used freshwater physics to match currently used surface_flux_sog calculation.
2. Completely changed the way the freshwater effects on nitrate, silicon etc
occur.  We are now considering that the physical circulation causes a constant
advection of a background gradient.  The background gradient is provided by
the difference between river concentrations and deep concentrations.  It
is independent of the current variable concentrations.. except...

3. Biology uses these quantities and when it uses them nearly to exhaustion
it removes the background gradient and thus the advective flux.

(Of course, does not diff clean).