Renamed the overall diffusivity variable from nK to K because the old K variable is gone now, so there's no more conflict over the name.
Cleaned up top of file documentation.

Deleted nu from the list of temporary public variables because it is
no longer needed outside the module.

Added calc_turbulent_fluxes() subroutine to public subroutines list.
This subroutine preserves the wbar%* profiles calculation code that
has been refactored out of define_flux.f90.  It's here for potential
future use as a diagnostic aid but it isn't required for the model, so
the subroutine isn't called anywhere.

Did a lot of work on type definitions that are private to the module
to facilitate the on refactoring of turbulence code from SOG.f90 to
this module.  See the comments for details.

Changed numerical constants to be explicitly double precision,
generally by adding "d0" or "0d0".

Added a new shear diffusivity smoothing parameter that is read from
the infile, and used to calculate several other smoothing parameter
values that untimately are used to improve the numerics of the shear
diffusivity term profile.

Added the arguments h_i, and h_g to the calc_KPP_diffusivity()
subroutine.  These are the indices of the grid layer interface (h%i),
and grid point (h%g) immediately below the mixing layer depth.  They
are required for turbulence calculations, and components of the h%*
variable must be passed in.  They cannot be used from the mixing_layer
module here because that causes a circular module dependency.

Added explicit setting of the nu%*%total(0) value to zero to enforce
the boundary condition of no diffusion across the air-water boundary.
The boundary values are needed for an edge-case in the interior to
mixing layer diffusion matching code.

Deleted the wind_driven_turbulence() and
convection_driven_turbulence() subroutines.  They have been replaced
by a w_scale() function that returns the value of the appropriate
turbulent vertical velocity scale (w%*) at the specified depth.  This
eliminates the need to calculate profiles of w%* (though they still
exist as refactoring bridge code) and facilitates calculation of
K_ML%* at arbitrary depths with less interpolation (and no double
interpolation).

Refactored code for non-dimensional veritical diffusivity shape
functions (G(sigma)) into this module from SOG.f90,
shape_parameters.f90, interior_match.f90, and interior_match2.f90.

Refactored code for calculation of mixing layer diffusivity (K_ML%*),
modification of its value just above the mixing layer depth, and
setting of the overall diffusivity profile (K%*) into this module from
SOG.f90, modify_k.f90, and define_flux.f90.

The mixing layer diffusivity is the larger of the mixing layer
diffusivity and the shear diffusivity.  This handles the situation of,
for instance, the wind dying so the surface forcing produces minimal
mixing layer diffusivity, but significant shear remains in the mixing
layer.

Added code to the shear diffusion subroutine to smooth the shear
diffusivity over adjacent grid layer interfaces because the estimation
of shear diffusivity is noisy as it is the ratio of a pair of
differences.