New baroclinic pressure gradient subroutine.
1) now grid spacing independent (can handle variable grids)

2) pressure calculation now explicitly on the grid points

3) isopycnals popping through layers now handled correctly

4) size of domain corrected

5) decayscale now explicitly linked to keeping tilt of the isopycnals
under control (ie less than about 6 m elevation)

6) symmetry maintained by calculating both sides of domain rather than
subtracting one side from the centre

Notes:

1) this method of handling the baroclinic pressure gradients does reduces the
magnitude of the near-inertial mode (as is appropriate) and increase its
frequency (by so little for the SoG its hard to see.. fine if you make it
small).

2) this method does produce a basin mode but the dynamics are a little
different from the Kelvin wave and the period is about a factor of 2 too large.