Worked through this module trying to eliminate generation of NaN's (ie make it more robust).
1. Taking the cube-root of a negative number (well defined in theory
but pgf90 gets NaN... so do it explicitly with an if-loop in the
calculation of w-star

2. Added epsilon's to a number of divisions.

3. Add if-statement so that G-functions and ML diffusivities are not
calculated in the case of no mixed layer (no wind and stable or no
buoyancy forcing).

4. Extended range for w-scale calculation.  was d< h and definitely
needs to include h but also the Richardson number calculation asks for
this value below the mixed layer (below is equal to mixed layer).

5. The sign of L_mo is very important in the calculatio of zeta_h... made
sure the sign of zeta_h is corrected even is L_mo is nearly infinite.