The g95 FORTRAN 90 compiler has now been installed on the ocean machines, so SOG can be compiled using pgf90 on coho, or g95 on any machine.
With this change the Makefile now looks for FCOMPILER = pgf90 to
decide to use the pgf90 compiler, otherwise it uses g95. Compared to
the previous checking of HOST, this make it more obvious what's going
on, and paves the way for the buildbot-based testing framework.

If you are logged on to coho and want to build with pgf90, the command
would be:

$ make FCOMPILER=pgf90

To use g95 (on any machine):

$ make

or

$ make FCOMPILER=g95

if you want to be explicit.

If you're logged on to a machine other than coho, you can do the pgf90
compile with:

$ ssh coho "cd `pwd`; make FCOMPILER=pgf90"

The pwd in backticks expands to the present working directory on the
local machine before the command string in double quotes is sent to
coho.

It's probably possible to compile the code partly with pgf90 and
partly with g95, but I doubt that it will link, or run.  If you run
into that, do a make clean and build from scratch with the compiler
you want to use.