Added code to handle no-wind, cooling
When there is no wind, but cooling (and of course, in SOG, freshening,
the KPP code estimates the nonlocal mixing with a procedure that assumes
only cooling (Eqn 20).  It puts the total Buoyancy forcing (which can
be nearly zero) in the denominator but the temperature flux in the numerator.
To contain this, under these circumstances, we use only the temperature
buoyancy forcing to calculate the w* in the denominator.