Refactor northern return flow infuence code.
Done primarily to bring the "use northern return flow" question and
the parameters that it request together in the infile.
As a side-effect, the object dependency tree is cleaner.

**NOTE**
Legacy Fortran-style infiles have to be changed as a result of this work.
However, the change allows the YAML infile processor to handle the northern
return flow option and paves the way for handling other optional parameter
sets.